Finished User Routes

+++ routes/folders.js
@@ -20,7 +20,7 @@ router.get("/", verifyToken, async (req, res) => {
 
     // Find Child Folders
     let folders = await pool.query(
-      "SELECT * FROM folders WHERE folder_id = ($1);",
+      "SELECT folders.*, COUNT(notes.*) AS notes FROM folders LEFT JOIN notes ON folders.id = notes.folder_id WHERE folders.folder_id = ($1) GROUP BY folders.id;",
       [folder_id]
     );
 
@@ -40,6 +40,7 @@ router.post("/", verifyToken, async (req, res) => {
   try {
     let chan_id = req.user.chan_id;
     let { folder_id, title } = req.body;
+    if (!title) return res.status(400).send("Missing Parameters");
 
     if (!folder_id) {
       folder_id = await pool.query(
@@ -71,10 +72,10 @@ router.put("/", verifyToken, async (req, res) => {
     let folder_id = req.query.folder_id;
     let { title, parent_id } = req.body;
 
-    if (!folder_id || !parent_id)
-      if (!folder.rows.length) return res.status(400).send("Missing Data");
+    if (!folder_id || !title || !parent_id)
+      return res.status(400).send("Missing Data");
 
-    folder = await pool.query(
+    let folder = await pool.query(
       "UPDATE folders SET title = ($1), folder_id = ($2) WHERE id = ($3) AND chan_id = ($4) RETURNING *;",
       [title, parent_id, folder_id, chan_id]
     );
@@ -90,9 +91,9 @@ router.put("/", verifyToken, async (req, res) => {
 // Delete Folder w/ folder_id
 router.delete("/", verifyToken, async (req, res) => {
   try {
-    console.log("deleteFolder");
     let chan_id = req.user.chan_id;
     let folder_id = req.query.folder_id;
+    if (!folder_id) return res.status(400).send("Missing Data");
 
     await pool.query(
       "DELETE FROM folders WHERE id = ($1) AND chan_id = ($2);",
